+++ bootique-mvc/src/main/java/com/nhl/bootique/mvc/resolver/TemplateResolver.java
@@ -1,10 +1,8 @@
 package com.nhl.bootique.mvc.resolver;
 
-import javax.ws.rs.container.ResourceInfo;
-
 import com.nhl.bootique.mvc.Template;
 
 public interface TemplateResolver {
 
-	Template resolve(String templateName, ResourceInfo resourceInfo);
+	Template resolve(String templateName, Class<?> viewType);
 }