+++ RELEASE-NOTES.md
@@ -2,3 +2,4 @@
 
 * #1 Mustache Integration
 * #2 Better template resolving mechanism
+* #5 Allow to align template location with an arbitrary class.